026 (PR A) review round 3: nine findings applied, three-dataset evaluation
Oracle checker names its file; a cache hit requires a complete float sidecar; the study rounds half away from zero in f64; vector(id) returns Result and refuses a damaged row as the scan does; rescored_by_sha256 asserted in both fixture-validity tests; scale and norm decoded once per row; hybrid baseline no longer materialises rows as an existence check; one zero-norm check in validate_query.
NFCorpus and FiQA evaluated in this pull request as Rule 5 requires: every delta inside 0.005, none beyond 0.001; engine-row agreement 0.9947-0.9956 on all three datasets.

+++ reference/int8_vectors_study.py
@@ -81,8 +81,10 @@ def quantise(vectors: np.ndarray) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
     peak = np.abs(vectors).max(axis=1, keepdims=True)
     scale = np.maximum(peak / np.float32(127.0), F32_MIN_POSITIVE).astype(np.float32)
     scale = np.where(peak > 0, scale, np.float32(1.0)).astype(np.float32)
-    q = (vectors / scale).astype(np.float32)
-    codes = (np.floor(np.abs(q) + np.float32(0.5)) * np.sign(q)).clip(-127, 127).astype(np.int8)
+    q = (vectors / scale).astype(np.float32).astype(np.float64)
+    # `f32::round` is half away from zero: the `+ 0.5` and the floor happen in f64, where adding
+    # 0.5 to an f32 value is exact; in f32 the add itself could round (0.49999997 + 0.5 → 1.0).
+    codes = (np.floor(np.abs(q) + 0.5) * np.sign(q)).clip(-127, 127).astype(np.int8)
     norms = np.sqrt((codes.astype(np.int64) ** 2).sum(axis=1)).astype(np.float64) * scale[:, 0].astype(np.float64)
     return codes, scale[:, 0], norms
